playground/disaster-routing-web/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==14.1
google-genai>=1.0.0
python-dotenv==1.0.1
